pyproject.toml
[project]
name = "instructvault"
version = "0.2.0"
description = "Git-first prompt registry + CI evals + lightweight runtime SDK (ivault)."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name="InstructVault OSS Maintainers"}]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
  "mypy>=1.10",
  "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/05satyam/instruct_vault"
Repository = "https://github.com/05satyam/instruct_vault"

[project.scripts]
ivault = "instructvault.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","B","UP","SIM","RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
